Breaking up functions into own directories.
